✨ feat: user_router.py GET user/me api 할시 riot_accounts도 List로 반환하도록 수정  -> 프론트단에서 방 생성시에 GET user/me 로 연동된 라이엇 계정을 드롭다운으로 나타내주기 위함
🗑️ remove: 필요없는 chat_room_dto.py 삭제

✨ feat: 새로운 기능 추가
🐛 fix: 버그 수정
🎨 style: 코드 포맷팅, 스타일 변경 (기능 변경 없음)
♻️ refactor: 코드 리팩터링
✅ test: 테스트 코드 추가 또는 수정
🔨 chore: 빌드 관련 설정 변경 또는 보조 작업
🚀 deploy: 배포 관련 작업
🗑️ remove: 파일 또는 코드 삭제
